requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
appdirs==1.4.4
argh==0.30.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
black==23.12.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.3
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.7
django-appconf==1.0.6
django-autoslug==1.9.8
django-celery-email==3.0.0
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==21.0.0
flake8==3.9.2
flower==2.0.1
frozenlist==1.4.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
mccabe==0.6.1
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
phonenumbers==8.12.33
Pillow==8.3.2
platformdirs==4.1.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.12
yarl==1.9.4
